+++ functions/public/Get-PSWorkItem.ps1
@@ -85,6 +85,7 @@ Function Get-PSWorkItem {
         $PSDefaultParameterValues['_verbose:block'] = 'Begin'
         _verbose -message $strings.Starting
         _verbose -message ($strings.PSVersion -f $PSVersionTable.PSVersion)
+        _verbose -message ($strings.UsingModule -f (Get-Command -name $MyInvocation.MyCommand).Version)
         _verbose -message ($strings.UsingDB -f $path)
         _verbose ($strings.DetectedCulture -f (Get-Culture))
     } #begin