+++ functions/public/Get-PSWorkItemPreferences.ps1
@@ -8,6 +8,7 @@ Function Get-PSWorkItemPreference {
         $PSDefaultParameterValues['_verbose:block'] = 'Begin'
         _verbose -message $strings.Starting
         _verbose -message ($strings.PSVersion -f $PSVersionTable.PSVersion)
+        _verbose -message ($strings.UsingModule -f (Get-Command -name $MyInvocation.MyCommand).Version)
         $PrefPath = Join-Path -Path $HOME -ChildPath .psworkitempref.json
     } #begin