+++ functions/public/Get-PSWorkItemReport.ps1
@@ -17,6 +17,7 @@ Function Get-PSWorkItemReport {
         $PSDefaultParameterValues["_verbose:block"] = "Begin"
         _verbose -message $strings.Starting
         _verbose -message ($strings.PSVersion -f $PSVersionTable.PSVersion)
+        _verbose -message ($strings.UsingModule -f (Get-Command -name $MyInvocation.MyCommand).Version)
         _verbose -message ($strings.UsingDB -f $path)
     } #begin