+++ functions/public/Get-PSWorkitemCategory.ps1
@@ -59,6 +59,7 @@ Function Get-PSWorkItemCategory {
         $PSDefaultParameterValues['_verbose:block'] = 'Begin'
         _verbose -message $strings.Starting
         _verbose -message ($strings.PSVersion -f $PSVersionTable.PSVersion)
+        _verbose -message ($strings.UsingModule -f (Get-Command -name $MyInvocation.MyCommand).Version)
         _verbose -message ($strings.UsingDB -f $path)
         Write-Debug 'Using bound parameters'
         $PSBoundParameters | Out-String | Write-Debug