+++ functions/public/New-PSWorkItem.ps1
@@ -99,6 +99,7 @@ Function New-PSWorkItem {
         $PSDefaultParameterValues['_verbose:block'] = 'Begin'
         _verbose -message $strings.Starting
         _verbose -message ($strings.PSVersion -f $PSVersionTable.PSVersion)
+        _verbose -message ($strings.UsingModule -f (Get-Command -name $MyInvocation.MyCommand).Version)
         _verbose -message ($strings.UsingDB -f $path)
         Try {
             $conn = Open-MySQLiteDB -Path $Path -ErrorAction Stop