+++ functions/public/Update-PSWorkItemPreference.ps1
@@ -27,6 +27,7 @@ Function Update-PSWorkItemPreference {
         $PSDefaultParameterValues["_verbose:block"] = "Begin"
         _verbose -message $strings.Starting
         _verbose -message ($strings.PSVersion -f $PSVersionTable.PSVersion)
+        _verbose -message ($strings.UsingModule -f (Get-Command -name $MyInvocation.MyCommand).Version)
         $FilePath = Join-Path -Path $HOME -ChildPath ".psworkitempref.json"
     } #begin